+++ src/main/java/net/more_rpg_classes/effect/FrostedEffect.java
@@ -1,27 +1,40 @@
 package net.more_rpg_classes.effect;
 
+import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectCategory;
 import net.minecraft.entity.effect.StatusEffectInstance;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.registry.tag.EntityTypeTags;
 import net.more_rpg_classes.MRPGCMod;
+import net.spell_power.api.SpellSchools;
+
+import static net.more_rpg_classes.util.CustomMethods.spellSchoolDamageCalculation;
+import static net.more_rpg_classes.util.CustomMethods.stackFreezeStacks;
 
 public class FrostedEffect extends StatusEffect {
     public FrostedEffect(StatusEffectCategory statusEffectCategory, int color) {
-        super(
-                statusEffectCategory, color);
+        super(statusEffectCategory, color);
     }
 
-    public void onApplied(LivingEntity entity,  int amplifier) {
-        super.onApplied(entity, amplifier);
-        if(!entity.hasStatusEffect(MRPGCEffects.FROZEN_SOLID.registryEntry)){
-            entity.setFrozenTicks(entity.getFrozenTicks() + (60 * amplifier));
-            if(amplifier == MRPGCMod.effectsConfig.value.frosted_amplifier_frozen_solid_conversion){
-                entity.addStatusEffect(new StatusEffectInstance(MRPGCEffects.FROZEN_SOLID.registryEntry,100,0,false,false,true));
-                entity.removeStatusEffect(MRPGCEffects.FROSTED.registryEntry);
+    public void onApplied(LivingEntity livingEntity,  int amplifier) {
+        super.onApplied(livingEntity, amplifier);
+        spellSchoolDamageCalculation(SpellSchools.FROST,1.0F,livingEntity, (PlayerEntity) livingEntity.getLastAttacker());
+        EntityType<?> type = livingEntity.getType();
+        if(type.isIn(EntityTypeTags.FREEZE_IMMUNE_ENTITY_TYPES)) {
+            if(!livingEntity.hasStatusEffect(MRPGCEffects.FROZEN_SOLID.registryEntry)){
+                stackFreezeStacks(livingEntity,20);
+                if(amplifier == MRPGCMod.effectsConfig.value.frosted_amplifier_frozen_solid_conversion){
+                    livingEntity.addStatusEffect(new StatusEffectInstance(MRPGCEffects.FROZEN_SOLID.registryEntry,
+                            100,0,false,false,true));
+                    livingEntity.removeStatusEffect(MRPGCEffects.FROSTED.registryEntry);
+                }
+            }else{
+                livingEntity.removeStatusEffect(MRPGCEffects.FROSTED.registryEntry);
             }
-        }else{
-            entity.removeStatusEffect(MRPGCEffects.FROSTED.registryEntry);
+        } else{
+            livingEntity.removeStatusEffect(MRPGCEffects.FROSTED.registryEntry);
         }
 
     }