+++ src/main/java/net/more_rpg_classes/effect/MRPGCEffects.java
@@ -86,10 +86,6 @@ public class MRPGCEffects {
 
         HealthImpacting.configureHealingTaken(GRIEVOUS_WOUNDS.effect,  MRPGCMod.effectsConfig.value.grievous_wounds_healing_taken);
 
-        OnRemoval.configure(FROZEN_SOLID.effect, (context) -> {
-            FrozenSolidEffect.onRemove(context.entity());
-        });
-
         for (var entry: entries) {
             entry.register();
         }